Split Transport + decompose Routing/OAuth/AsyncSequence helpers
file_length / type_body_length splits via per-concern extension files:
- SessionManager (634 → 57) → +Sessions / +Streams / +Broadcasting /
  +EventID.
- TCPBonjourTransport (566 → 248) → +Lifecycle / +Listener /
  +Connections / +Broadcasting.
- HTTPSSETransport (596 → 224) → +Authorize / +KeepAlive / +SSE /
  +Broadcasting / +Routes.

Function decompositions:
- HTTPSSETransport.authorize: 55-line cc-15 → 40 + checkJWERejection,
  authorizeForExistingSession.
- MCPRoutes.handleStreamableHTTP: 105/cc-18 → ~45 + 7 helpers.
- MCPRoutes.handleSSE: 90/cc-18 → ~15 + 8 helpers in MCPRoutes+SSE.
- OAuthRoutes.handleOAuthProxy: 156/cc-31 → ~30 + 6 helpers in
  OAuthRoutes+Proxy.
- OpenAPIRoutes.handleToolCallAsync: 102/cc-23 → ~35 + 5 helpers.
- OAuthConfigurationError.toOAuthConfiguration: 52 → ~30 +
  parseRequiredURL / parseOptionalURL.
- AsyncSequence+SSE.next: cc-11 → handleLine + applyField.
- HTTPSSETransport.start: extracted bindingError helper.

Access widening from private → internal (for cross-file refs):
SessionManager: BufferedEvent, StreamRecord, streams, sessionStreams,
primaryGeneralStreamIDs, eventBufferCapacity, destroySession,
updateSessionExpiry, cleanupExpiredState, removeStream,
selectPrimaryGeneralStream, sendSSE, sendPrimingEvent, parseEventID,
makeEventID. TCPBonjourTransport: queue, state, maxRetryDelay,
TransportState, createListener, installStateHandler,
handleListenerState, isRetryableError, scheduleRetry,
handleNewConnection, cleanupConnection, startReceiveLoop, receiveNext,
handleLine; port setter → internal(set). HTTPSSETransport: port setter
→ internal(set). OAuthRoutes: getBaseURL.

Build clean. All 415 tests pass.
