+++ Sources/SwiftMCP/Transport/TCPBonjourTransport.swift
@@ -26,18 +26,18 @@ public final class TCPBonjourTransport: Transport, @unchecked Sendable {
     public let serviceDomain: String
     public let acceptLocalOnly: Bool
     public let preferIPv4: Bool
-    public private(set) var port: UInt16?
+    public internal(set) var port: UInt16?
 
-    private let queue = DispatchQueue(label: "com.cocoanetics.SwiftMCP.TCPBonjourTransport")
-    private let state = TransportState()
+    internal let queue = DispatchQueue(label: "com.cocoanetics.SwiftMCP.TCPBonjourTransport")
+    internal let state = TransportState()
     internal lazy var sessionManager = SessionManager(transport: self)
 
     /// Maximum delay between retry attempts (in seconds).
-    private static let maxRetryDelay: UInt64 = 60
+    internal static let maxRetryDelay: UInt64 = 60
 
     // MARK: - Transport State
 
-    private actor TransportState {
+    internal actor TransportState {
         private(set) var isRunning: Bool = false
         private(set) var generation: UInt64 = 0
         private var listener: NWListener?
@@ -172,324 +172,6 @@ public final class TCPBonjourTransport: Transport, @unchecked Sendable {
             acceptLocalOnly: true,
             preferIPv4: true
         )
-    }
-
-    // MARK: - Lifecycle
-
-    public func start() async throws {
-        if await state.running() {
-            return
-        }
-
-        let listener = try createListener()
-        let generation = await state.start(listener: listener)
-        installStateHandler(on: listener, generation: generation)
-        listener.start(queue: queue)
-    }
-
-    public func run() async throws {
-        try await start()
-        await state.waitUntilStopped()
-    }
-
-    public func stop() async throws {
-        await state.stop()
-    }
-
-    /// Broadcasts a log message to all connected sessions.
-    public func broadcastLog(_ message: LogMessage) async {
-        await sessionManager.broadcastLog(message)
-    }
-
-    /// Broadcasts a tools list-changed notification to all connected sessions.
-    public func broadcastToolsListChanged() async {
-        await sessionManager.broadcastToolsListChanged()
-    }
-
-    /// Broadcasts a resources list-changed notification to all connected sessions.
-    public func broadcastResourcesListChanged() async {
-        await sessionManager.broadcastResourcesListChanged()
-    }
-
-    /// Broadcasts a prompts list-changed notification to all connected sessions.
-    public func broadcastPromptsListChanged() async {
-        await sessionManager.broadcastPromptsListChanged()
-    }
-
-    /// Broadcasts a resource-updated notification to all connected sessions.
-    /// - Parameter uri: The URI of the resource that was updated.
-    public func broadcastResourceUpdated(uri: URL) async {
-        await sessionManager.broadcastResourceUpdated(uri: uri)
-    }
-
-    // MARK: - Listener Creation
-
-    /// Creates a new NWListener with the transport's configuration.
-    private func createListener() throws -> NWListener {
-        let parameters = NWParameters.tcp
-        parameters.acceptLocalOnly = acceptLocalOnly
-        parameters.includePeerToPeer = false
-        if preferIPv4,
-           let ipOptions = parameters.defaultProtocolStack.internetProtocol as? NWProtocolIP.Options {
-            ipOptions.version = .v4
-        }
-
-        let listener: NWListener
-        if let port {
-            guard let nwPort = NWEndpoint.Port(rawValue: port) else {
-                throw TransportError.bindingFailed("Invalid TCP port: \(port)")
-            }
-            listener = try NWListener(using: parameters, on: nwPort)
-        } else {
-            listener = try NWListener(using: parameters)
-        }
-
-        let advertisedName = serviceName ?? server.serverName
-        listener.service = NWListener.Service(name: advertisedName, type: serviceType, domain: serviceDomain)
-        listener.newConnectionHandler = { [weak self] connection in
-            self?.handleNewConnection(connection)
-        }
-
-        return listener
-    }
-
-    /// Installs a state handler on a listener that captures only the generation token,
-    /// avoiding a strong reference to the listener itself.
-    private func installStateHandler(on listener: NWListener, generation: UInt64) {
-        listener.stateUpdateHandler = { [weak self] newState in
-            guard let self else { return }
-            Task {
-                await self.handleListenerState(newState, generation: generation)
-            }
-        }
-    }
-
-    // MARK: - Listener State Handling
-
-    private func handleListenerState(_ newState: NWListener.State, generation: UInt64) async {
-        switch newState {
-        case .ready:
-            if let boundPort = await state.listenerReady(generation: generation) {
-                port = boundPort
-            }
-            logger.info("TCP+Bonjour transport ready on port \(port.map(String.init) ?? "unknown")")
-
-        case .failed(let error):
-            if Self.isRetryableError(error) {
-                guard let delay = await state.listenerFailed(generation: generation) else {
-                    return  // stopped or stale generation
-                }
-                logger.warning("Bonjour listener failed (mDNSResponder unavailable): \(error). Retrying in \(delay)s.")
-                scheduleRetry(afterDelay: delay, failedGeneration: generation)
-            } else {
-                logger.error("TCP+Bonjour listener failed: \(error)")
-            }
-
-        case .cancelled:
-            logger.info("TCP+Bonjour listener cancelled")
-
-        default:
-            break
-        }
-    }
-
-    /// Returns `true` when the error indicates the mDNS daemon is unavailable
-    /// (DNS service error -65563 / `kDNSServiceErr_ServiceNotRunning`).
-    private static func isRetryableError(_ error: NWError) -> Bool {
-        if case .dns(let dnsError) = error, dnsError == -65563 {
-            return true
-        }
-        return false
-    }
-
-    // MARK: - Retry
-
-    /// Schedules a single retry attempt after the given delay.
-    /// The retry task is tracked in state and can be cancelled by `stop()`.
-    private func scheduleRetry(afterDelay delay: UInt64, failedGeneration: UInt64) {
-        let task = Task { [weak self] in
-            do {
-                try await Task.sleep(nanoseconds: delay * 1_000_000_000)
-            } catch {
-                return  // cancelled
-            }
-
-            guard let self else { return }
-
-            do {
-                let newListener = try self.createListener()
-                guard let newGeneration = await self.state.replaceListener(
-                    newListener,
-                    expectedGeneration: failedGeneration
-                ) else {
-                    // Transport was stopped or generation changed — discard
-                    newListener.cancel()
-                    return
-                }
-                self.installStateHandler(on: newListener, generation: newGeneration)
-                newListener.start(queue: self.queue)
-                self.logger.info("Bonjour listener re-created after retry, waiting for it to become ready.")
-            } catch {
-                // createListener() itself failed — schedule another retry
-                // with the same generation (state.listenerFailed already incremented attempt)
-                guard let nextDelay = await self.state.listenerFailed(generation: failedGeneration) else {
-                    return
-                }
-                self.logger.warning("Bonjour listener retry failed: \(error). Retrying in \(nextDelay)s.")
-                self.scheduleRetry(afterDelay: nextDelay, failedGeneration: failedGeneration)
-            }
-        }
-
-        Task {
-            await state.setRetryTask(task)
-        }
-    }
-
-    // MARK: - Send
-
-    public func send(_ data: Data) async throws {
-        guard let currentSession = Session.current else {
-            throw TransportError.bindingFailed("No active session for send")
-        }
-
-        guard let connection = await state.connection(for: currentSession.id) else {
-            throw TransportError.bindingFailed("TCP connection unavailable for session \(currentSession.id)")
-        }
-
-        let string = String(data: data, encoding: .utf8) ?? ""
-        logger.trace("TCP OUT:\n\n\(string)")
-
-        var out = data
-        out.append(Data("\n".utf8))
-
-        try await withCheckedThrowingContinuation { (continuation: CheckedContinuation<Void, Error>) in
-            connection.send(content: out, completion: .contentProcessed { error in
-                if let error {
-                    continuation.resume(throwing: error)
-                } else {
-                    continuation.resume()
-                }
-            })
-        }
-    }
-
-    // MARK: - Connections
-
-    private func handleNewConnection(_ connection: NWConnection) {
-        let connectionID = UUID()
-
-        Task {
-            let session = await sessionManager.session(id: connectionID)
-
-            connection.stateUpdateHandler = { [weak self] state in
-                guard let self else { return }
-                switch state {
-                case .ready:
-                    self.logger.info("TCP connection ready: \(connectionID)")
-                case .failed(let error):
-                    self.logger.error("TCP connection failed (\(connectionID)): \(error)")
-                    Task {
-                        await self.cleanupConnection(id: connectionID)
-                    }
-                case .cancelled:
-                    Task {
-                        await self.cleanupConnection(id: connectionID)
-                    }
-                default:
-                    break
-                }
-            }
-
-            await state.addConnection(id: connectionID, connection: connection)
-            connection.start(queue: queue)
-            startReceiveLoop(connection: connection, session: session, connectionID: connectionID)
-        }
-    }
-
-    private func cleanupConnection(id: UUID) async {
-        await state.removeConnection(id: id)
-        await sessionManager.removeSession(id: id)
-    }
-
-    private func startReceiveLoop(connection: NWConnection, session: Session, connectionID: UUID) {
-        let lineBuffer = LineBuffer()
-        receiveNext(connection: connection, session: session, connectionID: connectionID, lineBuffer: lineBuffer)
-    }
-
-    private func receiveNext(
-        connection: NWConnection,
-        session: Session,
-        connectionID: UUID,
-        lineBuffer: LineBuffer
-    ) {
-        connection.receive(
-            minimumIncompleteLength: 1,
-            maximumLength: 64 * 1024
-        ) { [weak self] data, _, isComplete, error in
-            guard let self else { return }
-
-            if let data, !data.isEmpty {
-                Task {
-                    await lineBuffer.append(data)
-                    let lines = await lineBuffer.processLines()
-                    for line in lines {
-                        await self.handleLine(line, session: session)
-                    }
-                }
-            }
-
-            if let error {
-                self.logger.error("TCP receive error (\(connectionID)): \(error)")
-                Task {
-                    await self.cleanupConnection(id: connectionID)
-                }
-                return
-            }
-
-            if isComplete {
-                Task {
-                    if let remaining = await lineBuffer.getRemaining() {
-                        await self.handleLine(remaining, session: session)
-                    }
-                    await self.cleanupConnection(id: connectionID)
-                }
-                return
-            }
-
-            self.receiveNext(
-                connection: connection,
-                session: session,
-                connectionID: connectionID,
-                lineBuffer: lineBuffer
-            )
-        }
-    }
-
-    private func handleLine(_ line: String, session: Session) async {
-        let trimmed = line.trimmingCharacters(in: .whitespacesAndNewlines)
-        guard !trimmed.isEmpty,
-              let data = line.data(using: .utf8) else {
-            return
-        }
-
-        logger.trace("TCP IN:\n\n\(line)")
-
-        await session.work { _ in
-            do {
-                let messages = try JSONRPCMessage.decodeMessages(from: data)
-                if await SessionInitializationGate.shouldReject(messages, for: session) {
-                    logger.warning("Rejected TCP request before initialize (\(session.id))")
-                    try await send(SessionInitializationGate.rejectionResponses(for: messages))
-                    return
-                }
-
-                let responses = await server.processBatch(messages)
-                guard !responses.isEmpty else { return }
-                try await send(responses)
-            } catch {
-                logger.error("Error decoding TCP message: \(error)")
-            }
-        }
     }
 }
 #else
